ISSUE-338 :truck: Move: rankingPageTabs 폴더 이름 competitionPageTabs로 변경
