+++ src/pages/competition/competitionPageTabs/MyScore.jsx
