+++ src/pages/competition/competitionPageTabs/Score1VS1.jsx
